Add Mixed Charts Widget Component preset (Not finished yet)
