+++ src/server/modules/DashboardBuilder/ModuleDashboardBuilderServer.ts
@@ -1382,6 +1382,171 @@ export default class ModuleDashboardBuilderServer extends ModuleServerBase {
 
 
         DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Année'
+        }, 'VarMixedChartsWidgetOptionsComponent.dimension_custom_filter_segment_types.0.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Mois'
+        }, 'VarMixedChartsWidgetOptionsComponent.dimension_custom_filter_segment_types.1.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Jour'
+        }, 'VarMixedChartsWidgetOptionsComponent.dimension_custom_filter_segment_types.2.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Semaine'
+        }, 'VarMixedChartsWidgetOptionsComponent.dimension_custom_filter_segment_types.3.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Année glissante'
+        }, 'VarMixedChartsWidgetOptionsComponent.dimension_custom_filter_segment_types.4.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Heure'
+        }, 'VarMixedChartsWidgetOptionsComponent.dimension_custom_filter_segment_types.5.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Minute'
+        }, 'VarMixedChartsWidgetOptionsComponent.dimension_custom_filter_segment_types.6.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Seconde'
+        }, 'VarMixedChartsWidgetOptionsComponent.dimension_custom_filter_segment_types.7.___LABEL___'));
+
+
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Options du widget'
+        }, 'var_mixed_charts_widget_options_component.separator.widget_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Titre du widget'
+        }, 'var_mixed_charts_widget_options_component.widget_title.title_name_code_text.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Couleur du fond'
+        }, 'var_mixed_charts_widget_options_component.bg_color.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Options du graphique'
+        }, 'var_mixed_charts_widget_options_component.separator.chart_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Titre du graphique'
+        }, 'var_mixed_charts_widget_options_component.separator.chart_title_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Afficher le titre'
+        }, 'var_mixed_charts_widget_options_component.title_display.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Couleur du titre'
+        }, 'var_mixed_charts_widget_options_component.title_font_color.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Taille du titre'
+        }, 'var_mixed_charts_widget_options_component.title_font_size.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Padding du titre'
+        }, 'var_mixed_charts_widget_options_component.title_padding.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Options de la légende'
+        }, 'var_mixed_charts_widget_options_component.separator.chart_legend_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Afficher la légende'
+        }, 'var_mixed_charts_widget_options_component.legend_display.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Position de la légende'
+        }, 'var_mixed_charts_widget_options_component.legend_position.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Couleur de la légende'
+        }, 'var_mixed_charts_widget_options_component.legend_font_color.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Taille du texte de la légende'
+        }, 'var_mixed_charts_widget_options_component.legend_font_size.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Largeur des blocs de la légende'
+        }, 'var_mixed_charts_widget_options_component.legend_box_width.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Padding de la légende'
+        }, 'var_mixed_charts_widget_options_component.legend_padding.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Utiliser le style des points pour la légende à la place des blocs'
+        }, 'var_mixed_charts_widget_options_component.legend_use_point_style.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Options de rendu du graphique'
+        }, 'var_mixed_charts_widget_options_component.separator.chart_render_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Options des données'
+        }, 'var_mixed_charts_widget_options_component.separator.datas_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Options des dimensions'
+        }, 'var_mixed_charts_widget_options_component.separator.datas_dimension_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Utiliser une dimension de donnée, issue d\'un champ ou d\'un filtre date segmenté'
+        }, 'var_mixed_charts_widget_options_component.has_dimension.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'La dimension est un champ ?'
+        }, 'var_mixed_charts_widget_options_component.dimension_is_vo_field_ref.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Champ de la dimension'
+        }, 'var_mixed_charts_widget_options_component.dimension_vo_field_ref.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Segmentation de la dimension date'
+        }, 'var_mixed_charts_widget_options_component.dimension_custom_filter_segment_type.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Max. de valeurs pour la dimension choisie'
+        }, 'var_mixed_charts_widget_options_component.max_dimension_values.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Permet de définir un nombre max de résultats pris en compte pour le graphique sur la dimension proposée. Par exemple si on a sélectionné une année et qu\'on segmente au jour, on peut limiter aux 10 premiers jours en indiquant 10 ici.'
+        }, 'var_mixed_charts_widget_options_component.max_dimension_values.tooltip.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Trier la dimension par un champ'
+        }, 'var_mixed_charts_widget_options_component.sort_dimension_by_vo_field_ref.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Trier la dimension par ordre croissant ?'
+        }, 'var_mixed_charts_widget_options_component.sort_dimension_by_asc.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Options de filtrage des valeurs'
+        }, 'var_mixed_charts_widget_options_component.separator.datas_filter_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Filtre des valeurs'
+        }, 'var_mixed_charts_widget_options_component.widget_filter_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation(
+            { 'fr-fr': "Configuration de l'échelle sur l'axe des abscisse" },
+            'var_mixed_charts_widget_options_component.scale_options_x.___LABEL___'
+        ));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation(
+            { 'fr-fr': "Configuration de l'échelle sur l'axe des ordonnées" },
+            'var_mixed_charts_widget_options_component.scale_options_y.___LABEL___'
+        ));
+
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Configurer les variables à afficher'
+        }, 'var_charts_options_component.separator.var_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Ajouter une variable (ou graphique)'
+        }, 'var_charts_options_component.add_var_chart_options_button.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Supprimer la variable'
+        }, 'var_charts_options_component.remove_var_chart_options_button.___LABEL___'));
+
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Options de la variable'
+        }, 'var_chart_options_item_component.separator.var_options.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Nom de la variable principale'
+        }, 'var_chart_options_item_component.var_name.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Type de graphe'
+        }, 'var_chart_options_item_component.graphe_type.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Couleur de fond'
+        }, 'var_chart_options_item_component.bg_color.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Nom de la variable principale'
+        }, 'var_chart_options_item_component.var_title.title_name_code_text.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Couleur de bordure'
+        }, 'var_chart_options_item_component.border_color.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
+            'fr-fr': 'Epaisseur de la bordure'
+        }, 'var_chart_options_item_component.border_width.___LABEL___'));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation(
+            { 'fr-fr': "Configuration de l'échelle sur l'axe des abscisse" },
+            'var_chart_options_item_component.scale_options_x.___LABEL___'
+        ));
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation(
+            { 'fr-fr': "Configuration de l'échelle sur l'axe des ordonnées" },
+            'var_chart_options_item_component.scale_options_y.___LABEL___'
+        ));
+
+        DefaultTranslationManager.registerDefaultTranslation(new DefaultTranslation({
             'fr-fr': 'Segmentation'
         }, 'tstz_filter_options.segment_type.___LABEL___'));
