+++ src/shared/modules/DashboardBuilder/vos/DashboardWidgetVO.ts
@@ -33,7 +33,7 @@ export default class DashboardWidgetVO implements IDistantVOBase, IWeightedItem,
     public static WIDGET_NAME_pageswitch: string = 'pageswitch';
     public static WIDGET_NAME_varpiechart: string = 'varpiechart';
     public static WIDGET_NAME_varlinechart: string = 'varlinechart';
-    public static WIDGET_NAME_varmixedchart: string = 'varmixedchart';
+    public static WIDGET_NAME_varmixedcharts: string = 'varmixedcharts';
 
     public id: number;
     public _type: string = DashboardWidgetVO.API_TYPE_ID;
