+++ src/shared/modules/DashboardBuilder/vos/VarChartOptionsVO.ts
@@ -1,5 +1,4 @@
 
-import { Scale } from "chart.js";
 import AbstractVO from "../../VO/abstract/AbstractVO";
 
 /**
@@ -12,14 +11,12 @@ export default class VarChartOptionsVO extends AbstractVO {
     constructor(
         public var_id?: number,
 
-        public filter_custom_field_filters?: { [field_id: string]: string },
+        public custom_filter_names?: { [field_id: string]: string },
 
+        public type?: string,  // The type of the graph (line, bar, radar)
         public bg_color?: string,
         public border_color?: string,
         public border_width?: number,
-        public scale_options_x?: Partial<Scale>,
-        public scale_options_y?: Partial<Scale>,
-        public scale_options_r?: Partial<Scale>,
     ) {
         super();
     }
