+++ src/vuejsclient/ts/components/Var/components/mixed-chart/VarMixedChartComponent.pug
@@ -1,5 +1,4 @@
-template(v-if="current_mixed_charts_data && current_mixed_charts_options")
-    mixedchart(
-        :data="current_mixed_charts_data"
-        :options="current_mixed_charts_options"
-    )
+mixedchart(
+    :data="current_mixed_charts_data"
+    :options="current_mixed_charts_options"
+)
