+++ src/vuejsclient/ts/components/dashboard_builder/widgets/var_chart_options/VarChartsOptionsComponent.scss
@@ -1,4 +1,4 @@
-.var_chart_options_component {
+.var_charts_options_component {
     .options_separator_lvl1 {
         font-size: medium;
         margin-top: 1em;
