+++ src/vuejsclient/ts/components/dashboard_builder/widgets/var_chart_options/item/VarChartOptionsItemComponent.pug
@@ -26,13 +26,38 @@
                 )
                     .lang_selector_option(slot="noOptions") -
 
+        .graphe_type 
+            .input_label {{ label('var_chart_options_item_component.graphe_type') }}
+            .input_wrapper
+                multiselect.graphe_type_selector(
+                    :preselect-first="false"
+                    :multiple="false"
+                    v-model="type"
+                    :options="graphe_types"
+                    :allow-empty="false",
+
+                    :searchable="true"
+                    :internal-search="true"
+                    
+                    select-label=''
+                    select-group-label=''
+                    selected-label=''
+                    deselect-label=''
+                    deselect-group-label=''
+
+                    placeholder=''
+                    
+                    name="graphe_type_selector"
+                )
+                    .lang_selector_option(slot="noOptions") -
+
         .bg_color
             .explication {{ label("var_chart_options_item_component.bg_color") }}
-            color-picker(v-model="bg_color" @change='handle_bg_color_change')
+            color-picker(v-model="bg_color")
 
         .border_color
             .explication {{ label("var_chart_options_item_component.border_color") }}
-            color-picker(v-model="border_color" @change='handle_border_color_change')
+            color-picker(v-model="border_color")
 
         .border_width
             .input_label {{ label('var_chart_options_item_component.border_width') }}
@@ -41,33 +66,17 @@
                     v-model="border_width"
                     type="number"
                 )
-        
-        .scale_options_x 
-            .input_label {{ label('var_chart_options_item_component.scale_options_x') }}
-            .input_wrapper
-                Chartjsscaleoptionscomponent(
-                    @on_change="handle_scale_options_x_change"
-                    :options="scale_options_x"
-                )
-                
-        .scale_options_y 
-            .input_label {{ label('var_chart_options_item_component.scale_options_y') }}
-            .input_wrapper
-                Chartjsscaleoptionscomponent(
-                    @on_change="handle_scale_options_y_change"
-                    :options="scale_options_y"
-                )
 
-        template(v-if='fie_could_get_custom_filter && fields_that_could_get_custom_filter.length')
+        template(v-if='fields_that_could_get_custom_filter && fields_that_could_get_custom_filter.length')
             .fields_that_could_get_custom_filter
-                template(v-for='fieldlds_that_that_could_get_custom_filter in fields_that_could_get_custom_filter')
+                template(v-for='field_that_could_get_custom_filter in fields_that_could_get_custom_filter')
                     .field_that_could_get_custom_filter
                         .field_id {{ field_that_could_get_custom_filter }}
                         .custom_filter 
                             input(
-                                type="text"
                                 v-model="custom_filter_names[field_that_could_get_custom_filter]"
                                 :disabled="true"
+                                type="text"
                             )
                         .existing_custom_filters_list(v-if='get_custom_filters && get_custom_filters.length')
                             template(v-for='custom_filter in get_custom_filters')
