+++ src/vuejsclient/ts/components/dashboard_builder/widgets/var_line_chart_widget/VarLineChartWidgetComponent.ts
@@ -734,7 +734,6 @@ export default class VarLineChartWidgetComponent extends VueComponentBase {
 
     @Watch('widget_options', { immediate: true })
     private async onchange_widget_options() {
-
         await this.throttled_update_visible_options();
     }
