+++ src/vuejsclient/ts/components/dashboard_builder/widgets/var_mixed_charts_widget/VarMixedChartsWidgetComponent.scss
@@ -1,4 +1,4 @@
-.var_mixed_chart_widget_component {
+.var_mixed_charts_widget_component {
     .chart_wrapper {
         height: -webkit-fill-available;
