+++ src/vuejsclient/ts/components/dashboard_builder/widgets/var_mixed_charts_widget/icon/VarMixedChartsWidgetIconComponent.pug
@@ -1,4 +1,4 @@
 .var_widget_icon
     template(v-if='!widget')
     template(v-else)
-        i.fa-lg.fa-fw.fa-duotone.fa-chart-line(aria-hidden="true")
+        i.fa-lg.fa-fw.fa-duotone.fa-area-chart(aria-hidden="true")
