+++ src/vuejsclient/ts/components/dashboard_builder/widgets/var_mixed_charts_widget/icon/VarMixedChartsWidgetIconComponent.scss
