+++ src/vuejsclient/ts/components/dashboard_builder/widgets/var_mixed_charts_widget/icon/VarMixedChartsWidgetIconComponent.ts
@@ -3,13 +3,13 @@ import { Prop } from 'vue-property-decorator';
 import 'vue-slider-component/theme/default.css';
 import DashboardWidgetVO from '../../../../../../../shared/modules/DashboardBuilder/vos/DashboardWidgetVO';
 import VueComponentBase from '../../../../VueComponentBase';
-import './VarMixedChartWidgetIconComponent.scss';
+import './VarMixedChartsWidgetIconComponent.scss';
 
 @Component({
-    template: require('./VarMixedChartWidgetIconComponent.pug'),
+    template: require('./VarMixedChartsWidgetIconComponent.pug'),
     components: {}
 })
-export default class VarMixedChartWidgetIconComponent extends VueComponentBase {
+export default class VarMixedChartsWidgetIconComponent extends VueComponentBase {
 
     @Prop()
     private widget: DashboardWidgetVO;
