+++ src/vuejsclient/ts/components/dashboard_builder/widgets/var_mixed_charts_widget/options/VarMixedChartsWidgetOptionsComponent.scss
@@ -1,4 +1,4 @@
-.var_mixed_chart_widget_options_component {
+.var_mixed_charts_widget_options_component {
     .options_separator_lvl1 {
         font-size: medium;
         margin-top: 1em;
